Implement Merge by Schema
